+++ source/app/service-providers/menu/menu.win32.ts
@@ -21,6 +21,7 @@ import RecentDocumentsProvider from '@providers/recent-docs'
 import WindowProvider from '@providers/windows'
 import CommandProvider from '@providers/commands'
 import LogProvider from '@providers/log'
+import { zoomIn, zoomOut } from './font-zoom'
 
 export default function getMenu (
   logger: LogProvider,
@@ -31,6 +32,7 @@ export default function getMenu (
   getCheckboxState: (id: string, init: boolean) => boolean,
   setCheckboxState: (id: string, val: boolean) => void
 ): MenuItemConstructorOptions[] {
+  const useGuiZoom = config.get('system.zoomType') === 'gui'
   // While on macOS we can just drop the following menuItem into the menu, the
   // win32-menu is also being used on Linux. Therefore, we use as fallback the
   // default, but ...
@@ -496,19 +498,24 @@ export default function getMenu (
           id: 'menu.reset_zoom',
           label: trans('menu.reset_zoom'),
           accelerator: 'Ctrl+0',
-          role: 'resetZoom'
+          // NOTE: Since the base font-size can change, resetting the editor
+          // font size zoom level is not semantically meaningful
+          enabled: useGuiZoom,
+          role: (useGuiZoom) ? 'resetZoom' : undefined
         },
         {
           id: 'menu.zoom_in',
           label: trans('menu.zoom_in'),
           accelerator: 'Ctrl+Plus',
-          role: 'zoomIn'
+          role: (useGuiZoom) ? 'zoomIn' : undefined,
+          click: (useGuiZoom) ? undefined : zoomIn(config)
         },
         {
           id: 'menu.zoom_out',
           label: trans('menu.zoom_out'),
           accelerator: 'Ctrl+-',
-          role: 'zoomOut'
+          role: (useGuiZoom) ? 'zoomOut' : undefined,
+          click: (useGuiZoom) ? undefined : zoomOut(config)
         },
         {
           type: 'separator'