+++ source/types/main/config-provider.d.ts
@@ -139,6 +139,7 @@ interface ConfigOptions {
     avoidNewTabs: boolean
     iframeWhitelist: string[]
     checkForUpdates: boolean
+    zoomType: 'gui'|'editor'
   }
   checkForBeta: boolean
   uuid: string